extract utility binds, content transform fns
towards basic dynamics pt. 1 (#124)
